+++ cruhon/cli.py
@@ -17,7 +17,7 @@ import argparse
 from pathlib import Path
 
 
-CRUHON_VERSION = "1.4.0"
+CRUHON_VERSION = "1.5.0"
 
 BANNER = f"""
   \033[36m╔═══════════════════════════╗